# AVX-512 accelerated batch SHA256 implementation, related features and minor cleanups