--- a/src/ballet/sha256/Local.mk
+++ b/src/ballet/sha256/Local.mk
@@ -6,6 +6,9 @@ endif
 ifdef FD_HAS_AVX
 $(call add-objs,fd_sha256_batch_avx,fd_ballet)
 endif
+ifdef FD_HAS_AVX512
+$(call add-objs,fd_sha256_batch_avx512,fd_ballet)
+endif
 
 $(call make-unit-test,test_sha256,test_sha256,fd_ballet fd_util)
 $(call run-unit-test,test_sha256)