--- a/src/ballet/sha256/fd_sha256.h
+++ b/src/ballet/sha256/fd_sha256.h
@@ -157,6 +157,12 @@ FD_PROTOTYPES_END
 #define FD_SHA256_BATCH_ALIGN     ...
 #define FD_SHA256_BATCH_FOOTPRINT ...
 
+/* FD_SHA256_BATCH_MAX returns the batch size used under the hood.
+   Will be positive.  Users should not normally need use this for
+   anything. */
+
+#define FD_SHA256_BATCH_MAX       ...
+
 /* A fd_sha256_batch_t is an opaque handle for a set of SHA-256
    calculations. */
 
@@ -241,83 +247,21 @@ fd_sha256_batch_abort( fd_sha256_batch_t * batch );
 
 #endif
 
-#if FD_HAS_AVX /* AVX accelerated batching implementation */
+#ifndef FD_SHA256_BATCH_IMPL
+#if FD_HAS_AVX512
+#define FD_SHA256_BATCH_IMPL 2
+#elif FD_HAS_AVX
+#define FD_SHA256_BATCH_IMPL 1
+#else
+#define FD_SHA256_BATCH_IMPL 0
+#endif
+#endif
 
-#define FD_SHA256_BATCH_ALIGN     (128UL)
-#define FD_SHA256_BATCH_FOOTPRINT (256UL)
-
-/* This is exposed here to facilitate inlining various operations */
-
-#define FD_SHA256_PRIVATE_BATCH_MAX (8UL)
-
-struct __attribute__((aligned(FD_SHA256_BATCH_ALIGN))) fd_sha256_private_batch {
-  void const * data[ FD_SHA256_PRIVATE_BATCH_MAX ]; /* AVX aligned */
-  ulong        sz  [ FD_SHA256_PRIVATE_BATCH_MAX ]; /* AVX aligned */
-  void *       hash[ FD_SHA256_PRIVATE_BATCH_MAX ]; /* AVX aligned */
-  ulong        cnt;
-};
-
-typedef struct fd_sha256_private_batch fd_sha256_batch_t;
-
-FD_PROTOTYPES_BEGIN
-
-/* Internal use only */
-
-void
-fd_sha256_private_batch_avx( ulong          batch_cnt,    /* In [1,FD_SHA256_PRIVATE_BATCH_MAX] */
-                             void const *   batch_data,   /* Indexed [0,FD_SHA256_PRIVATE_BATCH_MAX), aligned 32,
-                                                             only [0,batch_cnt) used, essentially a msg_t const * const * */
-                             ulong const *  batch_sz,     /* Indexed [0,FD_SHA256_PRIVATE_BATCH_MAX), aligned 32,
-                                                             only [0,batch_cnt) used */
-                             void * const * batch_hash ); /* Indexed [0,FD_SHA256_PRIVATE_BATCH_MAX), aligned 32,
-                                                             only [0,batch_cnt) used */
-
-FD_FN_CONST static inline ulong fd_sha256_batch_align    ( void ) { return alignof(fd_sha256_batch_t); }
-FD_FN_CONST static inline ulong fd_sha256_batch_footprint( void ) { return sizeof (fd_sha256_batch_t); }
-
-static inline fd_sha256_batch_t *
-fd_sha256_batch_init( void * mem ) {
-  fd_sha256_batch_t * batch = (fd_sha256_batch_t *)mem;
-  batch->cnt = 0UL;
-  return batch;
-}
-
-static inline fd_sha256_batch_t *
-fd_sha256_batch_add( fd_sha256_batch_t * batch,
-                     void const *        data,
-                     ulong               sz,
-                     void *              hash ) {
-  ulong batch_cnt = batch->cnt;
-  batch->data[ batch_cnt ] = data;
-  batch->sz  [ batch_cnt ] = sz;
-  batch->hash[ batch_cnt ] = hash;
-  batch_cnt++;
-  if( FD_UNLIKELY( batch_cnt==FD_SHA256_PRIVATE_BATCH_MAX ) ) {
-    fd_sha256_private_batch_avx( batch_cnt, batch->data, batch->sz, batch->hash );
-    batch_cnt = 0UL;
-  }
-  batch->cnt = batch_cnt;
-  return batch;
-}
-
-static inline void *
-fd_sha256_batch_fini( fd_sha256_batch_t * batch ) {
-  ulong batch_cnt = batch->cnt;
-  if( FD_LIKELY( batch_cnt ) ) fd_sha256_private_batch_avx( batch_cnt, batch->data, batch->sz, batch->hash );
-  return (void *)batch;
-}
-
-static inline void *
-fd_sha256_batch_abort( fd_sha256_batch_t * batch ) {
-  return (void *)batch;
-}
-
-FD_PROTOTYPES_END
-
-#else /* Reference batching implementation */
+#if FD_SHA256_BATCH_IMPL==0 /* Reference batching implementation */
 
 #define FD_SHA256_BATCH_ALIGN     (1UL)
 #define FD_SHA256_BATCH_FOOTPRINT (1UL)
+#define FD_SHA256_BATCH_MAX       (1UL)
 
 typedef uchar fd_sha256_batch_t;
 
@@ -342,6 +286,152 @@ static inline void * fd_sha256_batch_abort( fd_sha256_batch_t * batch ) { return
 
 FD_PROTOTYPES_END
 
+#elif FD_SHA256_BATCH_IMPL==1 /* AVX accelerated batching implementation */
+
+#define FD_SHA256_BATCH_ALIGN     (128UL)
+#define FD_SHA256_BATCH_FOOTPRINT (256UL)
+#define FD_SHA256_BATCH_MAX       (8UL)
+
+/* This is exposed here to facilitate inlining various operations */
+
+struct __attribute__((aligned(FD_SHA256_BATCH_ALIGN))) fd_sha256_private_batch {
+  void const * data[ FD_SHA256_BATCH_MAX ]; /* AVX aligned */
+  ulong        sz  [ FD_SHA256_BATCH_MAX ]; /* AVX aligned */
+  void *       hash[ FD_SHA256_BATCH_MAX ]; /* AVX aligned */
+  ulong        cnt;
+};
+
+typedef struct fd_sha256_private_batch fd_sha256_batch_t;
+
+FD_PROTOTYPES_BEGIN
+
+/* Internal use only */
+
+void
+fd_sha256_private_batch_avx( ulong          batch_cnt,    /* In [1,FD_SHA256_BATCH_MAX] */
+                             void const *   batch_data,   /* Indexed [0,FD_SHA256_BATCH_MAX), aligned 32,
+                                                             only [0,batch_cnt) used, essentially a msg_t const * const * */
+                             ulong const *  batch_sz,     /* Indexed [0,FD_SHA256_BATCH_MAX), aligned 32,
+                                                             only [0,batch_cnt) used */
+                             void * const * batch_hash ); /* Indexed [0,FD_SHA256_BATCH_MAX), aligned 32,
+                                                             only [0,batch_cnt) used */
+
+FD_FN_CONST static inline ulong fd_sha256_batch_align    ( void ) { return alignof(fd_sha256_batch_t); }
+FD_FN_CONST static inline ulong fd_sha256_batch_footprint( void ) { return sizeof (fd_sha256_batch_t); }
+
+static inline fd_sha256_batch_t *
+fd_sha256_batch_init( void * mem ) {
+  fd_sha256_batch_t * batch = (fd_sha256_batch_t *)mem;
+  batch->cnt = 0UL;
+  return batch;
+}
+
+static inline fd_sha256_batch_t *
+fd_sha256_batch_add( fd_sha256_batch_t * batch,
+                     void const *        data,
+                     ulong               sz,
+                     void *              hash ) {
+  ulong batch_cnt = batch->cnt;
+  batch->data[ batch_cnt ] = data;
+  batch->sz  [ batch_cnt ] = sz;
+  batch->hash[ batch_cnt ] = hash;
+  batch_cnt++;
+  if( FD_UNLIKELY( batch_cnt==FD_SHA256_BATCH_MAX ) ) {
+    fd_sha256_private_batch_avx( batch_cnt, batch->data, batch->sz, batch->hash );
+    batch_cnt = 0UL;
+  }
+  batch->cnt = batch_cnt;
+  return batch;
+}
+
+static inline void *
+fd_sha256_batch_fini( fd_sha256_batch_t * batch ) {
+  ulong batch_cnt = batch->cnt;
+  if( FD_LIKELY( batch_cnt ) ) fd_sha256_private_batch_avx( batch_cnt, batch->data, batch->sz, batch->hash );
+  return (void *)batch;
+}
+
+static inline void *
+fd_sha256_batch_abort( fd_sha256_batch_t * batch ) {
+  return (void *)batch;
+}
+
+FD_PROTOTYPES_END
+
+#elif FD_SHA256_BATCH_IMPL==2 /* AVX-512 accelerated batching implementation */
+
+#define FD_SHA256_BATCH_ALIGN     (128UL)
+#define FD_SHA256_BATCH_FOOTPRINT (512UL)
+#define FD_SHA256_BATCH_MAX       (16UL)
+
+/* This is exposed here to facilitate inlining various operations */
+
+struct __attribute__((aligned(FD_SHA256_BATCH_ALIGN))) fd_sha256_private_batch {
+  void const * data[ FD_SHA256_BATCH_MAX ]; /* AVX aligned */
+  ulong        sz  [ FD_SHA256_BATCH_MAX ]; /* AVX aligned */
+  void *       hash[ FD_SHA256_BATCH_MAX ]; /* AVX aligned */
+  ulong        cnt;
+};
+
+typedef struct fd_sha256_private_batch fd_sha256_batch_t;
+
+FD_PROTOTYPES_BEGIN
+
+/* Internal use only */
+
+void
+fd_sha256_private_batch_avx512( ulong          batch_cnt,    /* In [1,FD_SHA256_BATCH_MAX] */
+                                void const *   batch_data,   /* Indexed [0,FD_SHA256_BATCH_MAX), aligned 32,
+                                                                only [0,batch_cnt) used, essentially a msg_t const * const * */
+                                ulong const *  batch_sz,     /* Indexed [0,FD_SHA256_BATCH_MAX), aligned 32,
+                                                                only [0,batch_cnt) used */
+                                void * const * batch_hash ); /* Indexed [0,FD_SHA256_BATCH_MAX), aligned 32,
+                                                                only [0,batch_cnt) used */
+
+FD_FN_CONST static inline ulong fd_sha256_batch_align    ( void ) { return alignof(fd_sha256_batch_t); }
+FD_FN_CONST static inline ulong fd_sha256_batch_footprint( void ) { return sizeof (fd_sha256_batch_t); }
+
+static inline fd_sha256_batch_t *
+fd_sha256_batch_init( void * mem ) {
+  fd_sha256_batch_t * batch = (fd_sha256_batch_t *)mem;
+  batch->cnt = 0UL;
+  return batch;
+}
+
+static inline fd_sha256_batch_t *
+fd_sha256_batch_add( fd_sha256_batch_t * batch,
+                     void const *        data,
+                     ulong               sz,
+                     void *              hash ) {
+  ulong batch_cnt = batch->cnt;
+  batch->data[ batch_cnt ] = data;
+  batch->sz  [ batch_cnt ] = sz;
+  batch->hash[ batch_cnt ] = hash;
+  batch_cnt++;
+  if( FD_UNLIKELY( batch_cnt==FD_SHA256_BATCH_MAX ) ) {
+    fd_sha256_private_batch_avx512( batch_cnt, batch->data, batch->sz, batch->hash );
+    batch_cnt = 0UL;
+  }
+  batch->cnt = batch_cnt;
+  return batch;
+}
+
+static inline void *
+fd_sha256_batch_fini( fd_sha256_batch_t * batch ) {
+  ulong batch_cnt = batch->cnt;
+  if( FD_LIKELY( batch_cnt ) ) fd_sha256_private_batch_avx512( batch_cnt, batch->data, batch->sz, batch->hash );
+  return (void *)batch;
+}
+
+static inline void *
+fd_sha256_batch_abort( fd_sha256_batch_t * batch ) {
+  return (void *)batch;
+}
+
+FD_PROTOTYPES_END
+
+#else
+#error "Unsupported FD_SHA256_BATCH_IMPL"
 #endif
 
 #endif /* HEADER_fd_src_ballet_sha256_fd_sha256_h */