… src/disco/shred/fd_shred_dest.h
… src/disco/shred/fd_shredder.h
… src/disco/shred/test_shredder.c
… src/tango/quic/fd_quic_conn.h
… src/util/fd_util_base.h
… src/util/simd/Local.mk
… src/util/simd/fd_avx512.h
… src/util/simd/fd_avx512_wwi.h
… src/util/simd/fd_avx512_wwl.h
… src/util/simd/fd_avx512_wwu.h
… src/util/simd/fd_avx512_wwv.h
… src/util/simd/test_avx512.h
… src/util/simd/test_avx512_16x32.c
… src/util/simd/test_avx512_8x64.c
… src/util/tmpl/fd_map_dynamic.c
… src/util/tmpl/fd_map_giant.c
… src/util/tmpl/fd_pool.c