As mentioned in PR #760, here is the accelerated SHA-256 version.  On a 2.3 GHz Icelake server core with turbo disabled (same one mentioned previously for AVX-512 accelerated SHA-512 and ED25519), quick and dirty benchmarks saw a large batch asymptotic Ethernet equivalent bandwidth processing rate of:
```
impl      | small msg | large msg
----------+-----------+----------
portable  |  1.8 Gbps |  1.4 Gbps
shani     |  6.4 Gbps |  7.0 Gbps
avx       |  9.4 Gbps |  8.7 Gbps * same asymptotic w shani
avx512    | 18.9 Gbps | 18.4 Gbps * same asymptotic w shani
```
As before, the small message size was 14 bytes (i.e. a maximum UDP payload of a minimal Ethernet frame holding a UDP/IP4/VLAN/Ethernet packet) and the large message size was 1472 bytes (i.e. a maximum UDP payload of a non-Jumbo 1500 byte MTU Ethernet frame holding a UDP/IP4/VLAN/Ethernet packet).  

This is based on and extends the AVX-512 support added recently for faster SHA-512 hashing and ED25519 signature verification.  Included here are lower level APIs necessary (e.g. 16-wide vector uint support) and some minor cleanups that ended up easier to include here than separate out into additional PRs (most notably, a function attribute linting pass).  At this point, there is very good coverage for accelerated 16-wide vector int / uint and 8-wide vector long / ulong operations.

This implementation is a turn-the-crank widened version of the existing AVX implementation with some additional streamlining to take advantage of AVX-512 support for vector lane masks and the SHA-256 header was also tweaked to make it easier to switch implementations at compile time and/or in individual translation units without needing to change build targets.

As per the "same asymptotic comment" above, for small enough batches, this will fall back on the serial SHA-NI implementation when available and faster.

That is, the batch implementations (i.e. GPU style data parallel using generic vector operations) are _much_ _faster_ asymptotically than the SHA-NI based implementation (i.e. take silicon away that could have been used for more cores and use it to try to overcome the "never met a sequential dependency I didn't like" design philosophy pervasive in crypto protocols).

While this might seem surprising, it's usually much easier to do N independent tasks that each take time T in parallel than it is to do one of the tasks in time T/N, even when you throw custom hardware at it.  This is a real world example the that compute-constrained serial mindset taught in schools and reinforced by languages and tooling is completely misaligned with our massively-parallel speed-of-light constrained reality.  (Better still would be to stop designing cryptographic protocols as though computers are just fast Turing tape machines and instead design them for machines buildable in a universe with a finite speed of light ... but we make do with what we have.)

TL;DR

AVX-512 accelerated batch SHA-256 is ~11x-12x faster than the existing portable implementation, ~2.6x-3.0x faster than the existing SHA-NI accelerated implementation and ~2.0x faster than the existing AVX accelerated batch implementation per core for large numbers of similar sized messages.  YMMV.